+++ src/render/data.rs
@@ -4,6 +4,7 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
+use ansi_term::Colour;
 use pulldown_cmark::CowStr;
 
 #[derive(Debug, PartialEq)]
@@ -11,6 +12,7 @@ pub struct Link<'a> {
     pub(crate) index: u16,
     pub(crate) target: CowStr<'a>,
     pub(crate) title: CowStr<'a>,
+    pub(crate) colour: Colour,
 }
 
 /// Data associated with rendering state.
@@ -29,13 +31,24 @@ pub struct StateData<'a> {
 }
 
 impl<'a> StateData<'a> {
-    pub(crate) fn add_link(mut self, target: CowStr<'a>, title: CowStr<'a>) -> (Self, u16) {
+    /// Add a pending link to the state data.
+    ///
+    /// `target` is the link target, and `title` the link title to show after the URL.
+    /// `colour` is the colour to use for foreground text to differentiate between
+    /// different types of links.
+    pub(crate) fn add_link(
+        mut self,
+        target: CowStr<'a>,
+        title: CowStr<'a>,
+        colour: Colour,
+    ) -> (Self, u16) {
         let index = self.next_link;
         self.next_link += 1;
         self.pending_links.push(Link {
             index,
             target,
             title,
+            colour,
         });
         (self, index)
     }